Fix batch writes
Switch to the regular (non-enhanced) client
to perform the batch writes when initializing
the table. That way, we don't conflict with
the VersionedRecordExtension.

We explicitly initialize the `version` attribute
to `1`, so that when we switch to the enhanced
client for the other REST endpoints, the
optimistic locking will work properly.